+++ packages/cli/src/commands/secrets.js
@@ -1,6 +1,5 @@
 import chalk from 'chalk';
 import table from 'text-table';
-import mri from 'mri';
 import ms from 'ms';
 import strlen from '../util/strlen.ts';
 import { handleError, error } from '../util/error';
@@ -9,8 +8,8 @@ import exit from '../util/exit';
 import getScope from '../util/get-scope.ts';
 import confirm from '../util/input/confirm';
 import getCommandFlags from '../util/get-command-flags';
-import getPrefixedFlags from '../util/get-prefixed-flags';
 import { packageName, getCommandName, logo } from '../util/pkg-name.ts';
+import getArgs from '../util/get-args';
 
 const help = () => {
   console.log(`
@@ -81,14 +80,12 @@ let subcommand;
 let nextTimestamp;
 
 const main = async client => {
-  argv = mri(client.argv.slice(2), {
-    boolean: ['help', 'debug', 'yes'],
-    alias: {
-      help: 'h',
-      debug: 'd',
-      yes: 'y',
-      next: 'N',
-    },
+  argv = getArgs(client.argv.slice(2), {
+    '--yes': Boolean,
+    '--next': Number,
+    '--test-warning': Boolean,
+    '-y': '--yes',
+    '-N': '--next',
   });
 
   argv._ = argv._.slice(1);
@@ -134,7 +131,8 @@ async function run({ output, contextName, currentTeam, client }) {
   const secrets = new NowSecrets({ client, currentTeam });
   const args = argv._.slice(1);
   const start = Date.now();
-  const { 'test-warning': testWarningFlag } = argv;
+  const { '--test-warning': testWarningFlag } = argv;
+
   const commandName = getCommandName('secret ' + subcommand);
 
   if (subcommand === 'ls' || subcommand === 'list') {
@@ -190,14 +188,7 @@ async function run({ output, contextName, currentTeam, client }) {
     }
 
     if (pagination && pagination.count === 20) {
-      const prefixedArgs = getPrefixedFlags(argv);
-      const flags = getCommandFlags(prefixedArgs, [
-        '_',
-        '--next',
-        '-N',
-        '-d',
-        '-y',
-      ]);
+      const flags = getCommandFlags(argv, ['_', '--next', '-N', '-d', '-y']);
       const nextCmd = `secrets ${subcommand}${flags} --next ${pagination.next}`;
       output.log(`To display the next page run ${getCommandName(nextCmd)}`);
     }
@@ -225,7 +216,7 @@ async function run({ output, contextName, currentTeam, client }) {
 
     if (theSecret) {
       const yes =
-        argv.yes ||
+        argv['--yes'] ||
         (await readConfirmation(client, output, theSecret, contextName));
       if (!yes) {
         output.print(`Canceled. Secret not deleted.\n`);