+++ packages/node/src/types.ts
@@ -1,5 +1,5 @@
 import { ServerResponse, IncomingMessage } from 'http';
-import type { Headers } from 'node-fetch';
+import type { Headers } from 'undici';
 
 export type VercelRequestCookies = { [key: string]: string };
 export type VercelRequestQuery = { [key: string]: string | string[] };
@@ -44,6 +44,6 @@ export type NowApiHandler = VercelApiHandler;
 export interface VercelProxyResponse {
   status: number;
   headers: Headers;
-  body: Buffer | NodeJS.ReadableStream;
+  body: ReadableStream<Uint8Array> | Buffer | null;
   encoding: BufferEncoding;
 }