+++ packages/remix/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/remix-builder
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`9e3827c78`](https://github.com/vercel/vercel/commit/9e3827c785e1bc45f2bed421132167381481770f)]:
+  - @vercel/build-utils@7.1.0
+
+## 2.0.1
+
+### Patch Changes
+
+- Update `@remix-run/dev` fork to v1.19.3 ([#10381](https://github.com/vercel/vercel/pull/10381))
+
 ## 2.0.0
 
 ### Major Changes