+++ packages/static-build/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/static-build
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes